requirements.txt
Django==5.2
google-auth-oauthlib==1.2.2
google-api-python-client==2.169.0